apps/backend/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
supabase==2.11.0
openai==1.58.1
